backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.135.2",
    "langchain-core>=1.2.23",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.3",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "sse-starlette>=3.3.3",
    "tavily-python>=0.7.23",
    "uvicorn[standard]>=0.42.0",
    "yfinance>=1.2.0",
]
